+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release 5.3.0:
    - In `CredentialResponseParameters`, change type of `credential` from `String` to `JsonElement`
    - Support issuing credentials for multiple proofs
    - For `CredentialRequestInput` add subclass `AuthorizationDetails` to request only some specific claims
+   - In `CredentialRequestParameters` change type of `claims` to `JsonElement`, to support ISO mDoc and SD-JWT claims (cf. `SupportedCredentialFormat`)
 
 Release 5.2.0:
 - New `Initializer` object in `vck-openid` which needs to be called at the start of the project if artifact is used